flask/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
Brotli==1.0.7
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
fake-useragent==0.1.11
Flask==1.1.2
Flask-Compress==1.5.0
Flask-HTMLmin==2.0.0
Flask-Limiter==1.3.1
flask-marshmallow==0.12.0
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
htmlmin==0.1.12
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jsmin==2.2.2
lesscpy==0.14.0
limits==1.5.1
lxml==4.5.1
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.23.0
numpy==1.18.4
parse==1.15.0
persiantools==1.4.1
ply==3.11
pycountry==19.8.18
pyee==7.0.2
pyppeteer==0.2.2
pyquery==1.4.1
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2020.1
requests==2.23.0
requests-html==0.10.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
text-unidecode==1.3
tqdm==4.46.0
Unidecode==1.1.1
uWSGI==2.0.19.1
urllib3==1.25.9
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
xxhash==1.4.3
